refactor(bridge-flows): fix correctness + extract shared helpers
Correctness fixes from review:
- routeLabel lexical sort: compareRoute now sorts numerically on
  (sourceChainId, destChainId). Previously "143-…" lex-preceded "42220-…"
  because "1" < "4", giving wrong Monad/Celo mix order.
- Unify sort patterns through compareNullable; empty-string senders now
  sink rather than sort as valid values.
- Age → Time full rename: sort key renamed from "age" to "time" for
  consistency with the column header.
- Drop silent `?? "celo-mainnet"` fallback in TokenCell/RouteCell — was a
  correctness hazard the moment a non-Celo chain is added to the indexer
  but not to PROD_NETWORK_BY_CHAIN_ID. Use networkForChainId everywhere.
- Remove unreachable `providerMessageId` truthiness guard and
  `pills.length === 0` fallback (schema makes both states impossible).
- Update misleading comment on usdValueAtSend — current state is
  "indexer-pricing not implemented", not "rarely fires".

Extractions for testability and reuse:
- src/lib/bridge-flows/sort.ts — sortTransfers, compareNullable, routeKey
- src/lib/bridge-flows/pricing.ts — transferAmountTokens,
  transferAmountUsd, usdPricedFromLiveRate
- src/lib/wormhole/urls.ts — wormholescanUrl (provider-namespaced)
- src/components/sortable-th.tsx — shared SortableTh<K>; removes
  duplicate from global-pools-table.tsx
- src/lib/networks.ts — exported networkForChainId helper; removes three
  local duplicates across the codebase
