+++ ui-dashboard/src/lib/wormhole/urls.ts
@@ -1,0 +1,8 @@
+/**
+ * Wormholescan is the canonical cross-chain trace UI for a Wormhole transfer,
+ * keyed by digest. Resolves even when the source tx hasn't been indexed by
+ * Wormholescan yet — useful for PENDING transfers.
+ */
+export function wormholescanUrl(digest: string): string {
+  return `https://wormholescan.io/#/tx/${digest}`;
+}
